+++ src/demo/ts/ephox/alloy/demo/forms/DemoRenders.ts
@@ -1,12 +1,11 @@
 import { FieldSchema, ValueSchema } from '@ephox/boulder';
-import { Fun } from '@ephox/katamari';
 import * as DomFactory from 'ephox/alloy/api/component/DomFactory';
 import * as ItemWidget from 'ephox/alloy/api/ui/ItemWidget';
 import { Menu } from 'ephox/alloy/api/ui/Menu';
 import { ToolbarGroup } from 'ephox/alloy/api/ui/ToolbarGroup';
+import { ItemSpec, NormalItemSpec, SeparatorItemSpec } from 'ephox/alloy/ui/types/ItemTypes';
 
-import { PremadeSpec, SimpleOrSketchSpec, AlloySpec, RawDomSchema } from 'ephox/alloy/api/component/SpecTypes';
-import { NormalItemSpec, ItemSpec, SeparatorItemSpec } from 'ephox/alloy/ui/types/ItemTypes';
+import { PartialMenuSpec } from '../../../../../../main/ts/ephox/alloy/ui/types/TieredMenuTypes';
 
 const demoItem = ValueSchema.objOf([
   FieldSchema.strictObjOf('data', [
@@ -158,7 +157,7 @@ const gridMenu = (menuSpec) => {
       Menu.parts().items({ })
     ],
     items: spec.items
-  };
+  } as PartialMenuSpec;
 };
 
 const menu = (menuSpec) => {