+++ src/main/ts/ephox/alloy/ui/types/InlineViewTypes.ts
@@ -1,11 +1,12 @@
 import { Option, Result } from '@ephox/katamari';
 
-import { Element } from '@ephox/sugar';
 import { AlloyBehaviourRecord } from '../../api/behaviour/Behaviour';
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import { SketchBehaviours } from '../../api/component/SketchBehaviours';
-import { AlloySpec, RawDomSchema, SketchSpec } from '../../api/component/SpecTypes';
+import { AlloySpec, RawDomSchema } from '../../api/component/SpecTypes';
 import { SingleSketch, SingleSketchDetail, SingleSketchSpec } from '../../api/ui/Sketcher';
+import { AnchorSpec } from '../../positioning/mode/Anchoring';
+import { TieredData, TieredMenuSpec } from './TieredMenuTypes';
 
 export interface InlineViewDetail extends SingleSketchDetail {
   uid: () => string;
@@ -17,6 +18,9 @@ export interface InlineViewDetail extends SingleSketchDetail {
   getRelated: () => (component: AlloyComponent) => Option<AlloyComponent>;
   lazySink: () =>  () => Result<AlloyComponent, Error>;
   eventOrder: () => Record<string, string[]>;
+  fireDismissalEventInstead: () => Option<{
+    event: () => string
+  }>;
 }
 
 export interface InlineViewSpec extends SingleSketchSpec {
@@ -29,19 +33,20 @@ export interface InlineViewSpec extends SingleSketchSpec {
   onHide?: (component: AlloyComponent) => void;
   getRelated?: (component: AlloyComponent) => Option<AlloyComponent>;
   eventOrder?: Record<string, string[]>;
+  fireDismissalEventInstead?: {
+    event?: string
+  };
 }
 
-// TYPIFY
-export interface InlineViewAnchor {
-  anchor: string;
-  x?: number;
-  y?: number;
-  item?: AlloyComponent;
-  root?: Element;
+export interface InlineMenuSpec {
+  data: TieredData,
+  menu: Partial<TieredMenuSpec>;
 }
 
 export interface InlineViewSketcher extends SingleSketch<InlineViewSpec, InlineViewDetail> {
-  showAt: (component: AlloyComponent, anchor: InlineViewAnchor, thing: AlloySpec) => void;
+  showAt: (component: AlloyComponent, anchor: AnchorSpec, thing: AlloySpec) => void;
+  showMenuAt: (component: AlloyComponent, anchor: AnchorSpec, menuSpec: InlineMenuSpec) => void;
   hide: (component: AlloyComponent) => void;
   isOpen: (component: AlloyComponent) => boolean;
+  getContent: (component: AlloyComponent) => Option<AlloyComponent>;
 }