+++ src/main/ts/ephox/alloy/ui/types/ItemTypes.ts
@@ -23,6 +23,8 @@ export interface WidgetItemSpec {
   data?: ItemDataTuple; // why is this necessary?
   dom: RawDomSchema;
   autofocus?: boolean;
+  widgetBehaviours?: AlloyBehaviourRecord;
+  ignoreFocus?: boolean;
   domModification?: DomModificationSpec;
 }
 
@@ -32,6 +34,8 @@ export interface WidgetItemDetail extends ItemDetail, CompositeSketchDetail {
   builder: () => <WidgetItemInfo>(buildInfo: WidgetItemInfo) => AlloySpec;
   autofocus: () => boolean;
   domModification: () => { };
+  widgetBehaviours: () => SketchBehaviours;
+  ignoreFocus: () => boolean;
   data: () => ItemDataTuple;
 }