+++ changelog.md
@@ -6,6 +6,77 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.29.0] - 2018-09-07
+### Added
+ - Docking.refresh() to recalculate the component's position and visibility
+
+## [3.28.0] - 2018-09-06
+### Added
+- Expanded the SlotContainer API to add:
+  - getSlotNames
+  - isShowing
+  - hideAllSlots
+### Fixed
+ - Sandbox cloaking no longer enforces position when no position attributes are applied.
+
+## [3.27.0] - 2018-09-06
+### Fixed
+ - Fixed Sliding behaviour responding to transitionend on nested elements
+ - Fixed types on Sliding behavior API
+
+## [3.26.0] - 2018-09-05
+### Added
+ - data alloy identifiers to the DOM nodes themselves. They are no longer in the
+ visible HTML
+
+## [3.25.0] - 2018-09-05
+### Added
+ - dynamic configuration of debugging modes
+
+## [3.24.0] - 2018-09-04
+### Added
+ - InlineView.showMenuAt() to special-case positioning for inline menus
+ - Sandboxing.openWhileCloaked() convenience method
+
+## [3.23.0] - 2018-08-31
+### Added
+ - eventOrder for Dropdowns
+ - extra debugging information
+
+## [3.22.0] - 2018-08-29
+### Added
+ - dragging behaviour flag for repositionTarget (defaults to true)
+ - dragging behaviour handler (onDrag)
+
+## [3.21.0] - 2018-08-29
+### Added
+ - onChoose event to Sliders
+
+## [3.20.0] - 2018-08-28
+### Added
+ - Replacing.replaceAt and Replacing.replaceBy
+
+## [3.19.0] - 2018-08-23
+### Added
+ - Tooltipping API access to hideAllExclusive, and tooltipComponents in config
+ - DomFactory.simple and DomFactory.dom for quick generation of basic AlloySpec objects
+ - InlineView API: getContent
+ - Readable state for Flatgrid Keying types
+ - Support for matrix-style menus
+ - Consistent definitions for itemBehaviours and widgetBehaviours
+ - IgnoreFocus capability for item widgets
+ - Exposing onChangeTab and onDismissTab through TabSectionTypes
+ - Chain methods for TestStore
+
+## [3.18.0] - 2018-08-20
+### Added
+ - selectClasses and selectAttributes to HtmlSelect sketcher
+
+## [3.17.0] - 2018-08-10
+### Added
+ - Configuration for InlineView: fireDismissalEventInstead
+ - SystemEvents.dismissRequested()
+
 ## [3.16.0] - 2018-08-08
 ### Added
 - Reflecting behaviour